Create basic UI with a place for chat and browser
Install CefSharp package that is used for navigating to web pages
Add RequestHandlerWT class which is inherited from CefSharp.RequestHandler class. RequestHandlerWT class is overrides OnBeforeBrowse method which helps us to get an information of the main requests of the browser and do some logic with this information
Add DisplayHandlerWT class so that it overrides OnFullscreenModeChange method which implements a logic of fullscreen browser surfing/video watching
Add BrowserSettings class which manages internal Cef settings

Install SimpleTCP package for easy-to-use data transfering among clients
Add ChatClient class which manages the client's side of data transferring through the Internet
Add ChatServer class which manages the server's side of data transfering through the Internet. This class is a host for all objects of the ChatClient class that will be connected to the server.
Add MessageWT struct which describes a message entity for the WatchTogether Chating.
Add MessageTypeWT enum which describes two types of messages used in the app: user messages and system messages. User messages are shown to the user. System messages are user for transfering a system information between client and server. System message, for example can be "Play video", "Pause", "Rewind video to x seconds" and etc.

Add BrowserCommandParser which must have a functionality to parse a system message text into an instance which is inherited from IBrowserCommand interface

Signed-off-by: PandrPi <pavlo.andrushko.me@gmail.com>